Display the warning when patterns not available
