+++ assets/admin/editor-wizard/patterns-list.js
@@ -6,6 +6,7 @@ import {
 	BlockPreview,
 } from '@wordpress/block-editor';
 import { useSelect } from '@wordpress/data';
+import { applyFilters } from '@wordpress/hooks';
 import { __ } from '@wordpress/i18n';
 import { ENTER, SPACE } from '@wordpress/keycodes';
 import { getBlockType, getBlockFromExample } from '@wordpress/blocks';
@@ -59,6 +60,30 @@ const withoutLessonActions = ( block ) =>
 	'sensei-lms/lesson-actions' !== block.name;
 
 /**
+ * No patterns warning component.
+ */
+const NoPatternsWarning = () => {
+	/**
+	 * Filters the warning message when no layouts are available.
+	 *
+	 * @param {string} message Default warning message.
+	 * @return {string} Filtered warning message.
+	 */
+	const warningMessage = applyFilters(
+		'sensei.editorWizard.noLayoutsWarning',
+		__( 'No layouts available for this theme.', 'sensei-lms' )
+	);
+
+	return (
+		<div className="sensei-patterns-list__warning">
+			<div className="sensei-patterns-list__warning-title">
+				{ warningMessage }
+			</div>
+		</div>
+	);
+};
+
+/**
  * Patterns list component.
  *
  * @param {Object}   props          Component props.
@@ -74,6 +99,10 @@ const PatternsList = ( { onChoose } ) => {
 				'sensei-lms/post-content'
 			),
 	} ) );
+
+	if ( ! patterns || patterns.length === 0 ) {
+		return <NoPatternsWarning />;
+	}
 
 	return (
 		<div
